+++ R/solve_POMDP.R
@@ -30,23 +30,27 @@ solve_POMDP <- function(
   parameter = NULL,
   verbose = FALSE) {
   
-  ### write model to file and create model if a filename is specified
-  file_prefix <- tempfile(pattern = "pomdp_")
-  pomdp_filename <- paste0(file_prefix, ".POMDP") 
+  converged <- NA
+   
+  methods <- c("grid", "enum", "twopass", "witness", "incprune") 
+  # Not implemented:  "linsup", "mcgs"
+  method <- match.arg(method, methods)
   
-  # write model
-  if(is.character(model)) {
+  # do we have a model POMDP file?
+  if(is.character(model)) 
     model <- structure(list(model = read_POMDP(model)), class = "POMDP")
-    writeLines(model$model$problem, con = pomdp_filename)
-  } else {
-    write_POMDP(model, pomdp_filename)
-  }
-
   
   if(is.null(horizon)) horizon <- model$model$horizon
+  # time-dependent POMDP?
+  if(!is.null(horizon) && length(horizon) > 1) 
+    return(.solve_POMDP_time_dependent(
+      model = model, horizon = horizon, discount = discount,
+      terminal_values = terminal_values, method = method,
+      parameter = parameter, verbose = verbose))
   if(is.null(horizon) || horizon < 1) horizon <- Inf 
   else horizon <- floor(horizon)
-   
+  
+  
   if(is.null(terminal_values)) terminal_values <- model$model$terminal_values
   if(!is.null(terminal_values) && 
       length(terminal_values) == 1 && terminal_values == 0) terminal_values <- NULL
@@ -56,15 +60,18 @@ solve_POMDP <- function(
     warning("No discount rate specified. Using .9!")
     discount <- .9
   }
-    
-  converged <- NA
-   
-  methods <- c("grid", "enum", "twopass", "witness", "incprune") 
-  # Not implemented:  "linsup", "mcgs"
-  method <- match.arg(method, methods)
-    
   
-  # write terminal values
+  ### temp file names
+  file_prefix <- tempfile(pattern = "pomdp_")
+  pomdp_filename <- paste0(file_prefix, ".POMDP") 
+  
+  # write model POMDP file
+  if(!is.null(model$model$problem)) 
+    writeLines(model$model$problem, con = pomdp_filename)
+  else
+    write_POMDP(model, pomdp_filename)
+  
+  # write terminal values file
   if(!is.null(terminal_values)) {
     if(!is.matrix(terminal_values)) terminal_values <- rbind(terminal_values)
     if(ncol(terminal_values) != length(model$model$states))
@@ -193,3 +200,89 @@ print.POMDP_solution <- function(x, ...) {
  print(unclass(x))
 }
 
+
+# solve time-dependent POMDP
+# we can have different transition_probs, observation_probs or rewards
+.solve_POMDP_time_dependent <- function(model, horizon = NULL, ..., verbose = FALSE) {
+    
+  if(verbose) cat("\n+++++++++ time-dependent POMDP +++++++++\n", sep = "")
+  
+  if(is.null(horizon)) horizon <- model$model$horizon
+  n <- length(horizon)
+  if(n < 2) return(solve_POMDP(model, horizon, ..., verbose = verbose))
+  
+  # check what is time dependent
+  do_trans <- FALSE
+  if(is.list(model$model$transition_prob) && 
+      is.list(model$model$transition_prob[[1]])) {
+    if(length(model$model$transition_prob) != n) 
+      stop("Not the right number of time-dependent transition probability matrices specified!")
+    do_trans <- TRUE
+  }
+  
+  do_obs <- FALSE
+  if(is.list(model$model$observation_prob) && 
+      is.list(model$model$observation_prob[[1]])) {
+    if(length(model$model$observation_prob) != n) 
+      stop("Not the right number of time-dependent observation probability matrices specified!")
+    do_obs <- TRUE
+  }
+  
+  do_reward <- FALSE
+  if(is.list(model$model$reward) && 
+      is.list(model$model$reward[[1]])) {
+    if(length(model$model$reward) != n) 
+      stop("Not the right number of time-dependent rewards specified!")
+    do_reward <- TRUE
+  }
+  
+  if(verbose) { 
+    if(do_trans) cat(" * Using time-dependent transition probabilities.\n")
+    if(do_obs) cat(" * Using time-dependent observation probabilities.\n")
+    if(do_reward) cat(" * Using time-dependent rewards.\n")
+  }
+  
+  # solve POMDPS in reverse order
+  s <- list()
+  m <- model
+  for(i in n:1) {
+    if(i < n) m$model$terminal_values <- prev_alpha
+    m$model$horizon <- model$model$horizon[i]
+    
+    # match names instead of index?
+    if(is.null(names(m$model$horizon))) take <- i
+    else take <- names(m$model$horizon)
+    
+    if(verbose) cat(
+      "\n++++++++++++++++++++++++++++++++++++++++\n", 
+      "Solving problem ", i, " (", take, ") with horizon ", m$model$horizon, 
+      "\n", sep = "")
+   
+    
+    if(do_trans) m$model$transition_prob <- model$model$transition_prob[[take]]
+    if(do_obs) m$model$observation_prob <- model$model$observation_prob[[take]]
+    if(do_reward) m$model$reward <- model$model$reward[[take]]
+   
+    if(verbose) {
+      if(do_trans) {
+        cat("Using transition probabilities:\n") 
+        print(m$model$transition_prob)
+        }
+    }
+    
+    s[[i]] <- solve_POMDP(m, ..., verbose = verbose)
+    prev_alpha <- s[[i]]$solution$alpha[[1]]
+  }
+  
+  # combine the results
+  pgs <- unlist(lapply(s, FUN = function(x) x$solution$pg), recursive = FALSE)
+  alphas <- unlist(lapply(s, FUN = function(x) x$solution$alpha), recursive = FALSE)
+  reward <- sum(sapply(s, FUN = function(x) x$solution$total_expected_reward))
+  
+  m <- structure(list(model = model$model, solution = s[[1]]$solution), class = "POMDP")
+  m$solution$pg <- pgs
+  m$solution$alpha <- alphas
+  m$solution$horizon <- sum(model$model$horizon)
+  
+  m
+}