+++ NEWS.md
@@ -6,6 +6,7 @@
 * solve_POMDP not looks at convergence.
 * solve_POMDP gained paramter terminal_values.
 * solve_POMDP gained parameter discount to overwrite the discount rate specified in the model.
+* solve_POMDP can now solve POMDPs with time-dependent transition probabilities, observation probabilities and reward structure. 
 * write_POMDP gained parameter digits.
 * added read_POMDP to read POMDP files. 
 * added policy graph visualization with vizNetwork.